feat(daemon): thread trigger into spawnSession and wire handoff handler (#1416)
Thread the TelemetryTrigger from runOnce through to spawnSession so
pull-mode tasks can be distinguished from cron/sessionstart tasks at
session construction time.

Three behavioral changes for pull-mode tasks:

1. isNonInteractive: false -- keeps ask_question in the tool schema so
   the durable handoff handler can persist questions for eventual reply

2. Handoff handler installed on elicitationRouter BEFORE sendMessage
   and uninstalled in the finally block -- the ask-question-gate's
   hasHandler() probe passes for pull tasks while cron ticks never
   inherit the handler

3. Cron and sessionstart tasks are unchanged (isNonInteractive: true,
   no handler)

Extract builtinPruneExecFile, resolveWorktreePruneRoot, and
daemonTraceLabel into scheduler.helpers.ts to offset the growth against
the filesize ratchet (scheduler.ts: 578 -> 571 LOC). Re-exported
through scheduler.ts so external import paths are unchanged.
